dialog skeleton and add page status fix
