+++ blocks/site-details/renderCampaignsAnalytics.js
@@ -4,6 +4,7 @@ import {
   EMAIL_WORKER_API, parseFragment, SCRIPT_API,
 } from '../../scripts/scripts.js';
 import renderSkeleton from '../../scripts/skeletons.js';
+import { createDialog } from '../../scripts/dialogs.js';
 
 const createAnalyticsTableContent = (campaignAnalyticsData, search) => {
   if (!campaignAnalyticsData) {
@@ -100,7 +101,7 @@ const createAnalyticsTableContent = (campaignAnalyticsData, search) => {
                 ${clone.outerHTML}    
             </div>
           `);
-            window.createDialog(content);
+            createDialog(content);
           };
         });
       },
@@ -135,12 +136,12 @@ export default async function renderCampaignsAnalytics({
 
   container.innerHTML = `
         <ul class="campaign-list" data-type="analytics">
-          <li><a class="button action secondary ${
+          <li><a class="button selector action secondary ${
   window.location.pathname.startsWith(`${pathname}/campaign-analytics/`) ? '' : 'is-selected'
 }" href="${pathname}/campaign-analytics">All emails</a></li>
           ${Object.keys(campaignsData)
     .map(
-      (campaignSlug) => `<li data-campaign="${campaignSlug}"><a class="button action secondary ${
+      (campaignSlug) => `<li data-campaign="${campaignSlug}"><a class="button selector action secondary ${
         window.location.pathname === `${pathname}/campaign-analytics/${campaignSlug}` ? 'is-selected' : ''
       }" href="${pathname}/campaign-analytics/${campaignSlug}">${campaignsData[campaignSlug].name}</li></a>`,
     )
