+++ styles/skeletons.css
@@ -1,0 +1,28 @@
+.skeleton {
+    position: relative;
+    overflow: hidden;
+    background-color: var(--light-gray);
+    border-radius: var(--small-border-radius);
+}
+
+.skeleton::before {
+    content: '';
+    position: absolute;
+    left: 0;
+    top: 0;
+    height: 100%;
+    width: 100%;
+    background: linear-gradient(90deg, transparent, rgba(0 0 0 / 5%), transparent);
+    transform: translateX(-100%);
+    animation: skeleton-slide 1s ease-in-out infinite;
+}
+
+@keyframes skeleton-slide {
+    0% {
+        transform: translateX(-100%);
+    }
+
+    100% {
+        transform: translateX(100%);
+    }
+}
