+++ styles/styles.css
@@ -49,10 +49,6 @@ html {
   scroll-behavior: smooth;
 }
 
-html:has(dialog[open]) {
-  overflow: hidden;
-}
-
 * {
   box-sizing: border-box;
   scroll-margin-top: var(--header-height);
@@ -95,7 +91,6 @@ main {
 }
 
 /* main of fragments */
-dialog main,
 main main {
   padding-top: 0;
 }
@@ -344,6 +339,10 @@ input[type="color"] {
   text-transform: none;
 }
 
+.button.action:hover {
+  background-color: var(--light-gray);
+}
+
 .button.destructive {
   background-color: var(--red);
   border-color: var(--red);
@@ -355,7 +354,15 @@ input[type="color"] {
   color: var(--black);
 }
 
+.button.selector {
+  border-color: var(--light-blue);
+}
 
+.button.selector.is-selected {
+  border-color: var(--light-blue);
+  background-color: var(--light-blue);
+  color: var(--white);
+}
 
 .button.loading {
   pointer-events: none;
@@ -404,237 +411,6 @@ table:has(tbody:empty) {
 .clusterize-scroll {
   max-height: calc(100vh / 2) !important;
   height: 100%;
-}
-
-/* MARK: Dialogs */
-dialog {
-  padding: 0;
-  border: none;
-  box-shadow: 0 0 20px 5px var(--gray);
-  border-radius: var(--border-radius);
-}
-
-dialog:has(.dialog-content) {
-  width: 800px;
-  max-width: 80vw;
-  max-height: 80vh;
-  overflow: visible;
-}
-
-dialog.fullscreen {
-  width: 90vw;
-  height: 90vh;
-  max-width: 90vw;
-  max-height: 90vh;
-}
-
-dialog .dialog-content {
-  padding: 16px;
-  height: 100%;
-  min-height: 400px;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  overflow: auto;
-  max-height: 80vh;
-}
-
-dialog.fullscreen .dialog-content {
-  max-height: 90vh;
-}
-
-dialog .dialog-content > div {
-  overflow: auto;
-  display: flex;
-  flex-direction: column;
-  flex: 1;
-}
-
-dialog .preview {
-  --preview-max-height: 300px;
-
-  height: auto;
-  max-height: var(--preview-max-height);
-  max-width: 100%;
-  overflow: hidden;
-  display: flex;
-  justify-content: center;
-}
-
-dialog .preview:empty {
-  display: none;
-}
-
-dialog .preview img {
-  display: block;
-  max-width: 128px;
-  width: 100%;
-  height: auto;
-  max-height: var(--preview-max-height);
-  object-fit: contain;
-  object-position: center;
-  filter: var(--icon-drop-shadow);
-}
-
-dialog .block-preview .block-iframe {
-  width: 100%;
-  height: 60vh;
-}
-
-dialog.loading {
-  pointer-events: none;
-}
-
-dialog.loading::before {
-  content: attr(data-loading-text);
-  font-size: 24px;
-  font-weight: 600;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  inset: 0;
-  z-index: 1;
-  background-color: var(--loading-backdrop);
-}
-
-dialog .dialog-content .dialog-button-container {
-  display: flex;
-  flex-flow: row wrap;
-  gap: 16px;
-  justify-content: flex-end;
-  padding-top: 16px;
-  overflow: visible;
-  flex: 0;
-}
-
-dialog .dialog-content .dialog-button-container:empty {
-  display: none;
-}
-
-dialog .dialog-content .dialog-button-container > * {
-  align-items: flex-start;
-}
-
-dialog .dialog-content > div:has(> .centered-info),
-dialog .centered-info {
-  display: flex;
-  flex: 1;
-  text-align: center;
-}
-
-dialog .centered-info {
-  align-items: center;
-  justify-content: center;
-}
-
-dialog .warning,
-dialog .warning * {
-  color: var(--orange);
-}
-
-dialog .error,
-dialog .error * {
-  color: var(--red);
-}
-
-@media (width < 768px) {
-  dialog {
-    min-width: unset;
-    max-width: none;
-    width: calc(100% - 32px);
-  }
-}
-
-dialog .close {
-  position: absolute;
-  top: -16px;
-  right: -16px;
-  border-radius: 50%;
-  height: 36px;
-  width: 36px;
-}
-
-dialog.alert-dialog {
-  width: 400px;
-}
-
-dialog.alert-dialog .dialog-content {
-  min-height: 200px;
-}
-
-dialog h3 {
-  text-align: center;
-}
-
-dialog form {
-  max-width: 400px;
-  width: 100%;
-}
-
-dialog form:last-child {
-  margin: auto;
-}
-
-dialog label {
-  display: block;
-  margin-bottom: 8px;
-  font-size: 14px;
-}
-
-dialog textarea {
-  min-height: 100px;
-  resize: none;
-  max-height: 400px;
-}
-
-dialog .columns {
-  display: flex;
-  gap: 32px;
-  flex-wrap: wrap;
-  height: 100%;
-  justify-content: center;
-}
-
-dialog .columns > * {
-  flex: 1;
-}
-
-dialog .columns iframe {
-  border: none;
-  height: auto;
-  max-width: 600px;
-  min-height: 400px;
-}
-
-@media (width < 768px) {
-  dialog .columns iframe {
-    display: none;
-  }
-}
-
-::backdrop {
-  background: rgba(0 0 0 / 20%);
-  backdrop-filter: blur(2px);
-}
-
-
-/* MARK: prev updates */
-
-dialog #revert-form ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-dialog #revert-form label {
-  display: flex;
-  gap: 20px;
-  justify-content: flex-start;
-}
-
-dialog #revert-form label input{
-  flex: 0;
 }
 
 /* MARK: headwire */
@@ -889,35 +665,6 @@ dialog #revert-form label input{
   line-height: 1.5;
   max-height: calc(1.5em * 3);
   overflow-y: auto;
-}
-
-.skeleton {
-  position: relative;
-  overflow: hidden;
-  background-color: var(--light-gray);
-  border-radius: var(--small-border-radius);
-}
-
-.skeleton::before {
-  content: '';
-  position: absolute;
-  left: 0;
-  top: 0;
-  height: 100%;
-  width: 100%;
-  background: linear-gradient(90deg, transparent, rgba(0 0 0 / 5%), transparent);
-  transform: translateX(-100%);
-  animation: skeleton-slide 1s ease-in-out infinite;
-}
-
-@keyframes skeleton-slide {
-  0% {
-    transform: translateX(-100%);
-  }
-
-  100% {
-    transform: translateX(100%);
-  }
 }
 
 .filter {
